+++ breath/lib/screens/calling/components/dialog_button.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class DialogButton extends StatelessWidget {
+  final String text;
+  final Color bgColor;
+  final Color textColor;
+  final VoidCallback onPressed;
+
+  DialogButton(this.text, this.bgColor, this.textColor, this.onPressed);
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      onPressed: onPressed,
+      style: ElevatedButton.styleFrom(
+        backgroundColor: bgColor,
+        minimumSize: Size(105, 40),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(14),
+        ),
+        padding: EdgeInsets.symmetric(horizontal: 20, vertical: 12),
+      ),
+      child: Text(
+        text,
+        style: TextStyle(
+            fontSize: 14, color: textColor, fontWeight: FontWeight.w500),
+      ),
+    );
+  }
+}
